pyproject.toml
[project]
name = "repl-benches"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-cpuinfo>=9.0.0",
]

[tool.uv.workspace]
members = [
    "provision/qemu_pin",
]

[project.optional-dependencies]
search = [
    "usearch==2.19.0",
    "faiss-cpu>=1.11.0",
    "annoy>=1.17.3",
    "seaborn>=0.13.2",
    "scikit-learn>=1.7.0",
    "h5py>=3.14.0",
    "requests>=2.32.4",
]
